Refactor keymap config into modular hooks and schema constants
