--- src/constants/defaults.ts
@@ -1,4 +1,0 @@
-export const DEFAULT_HOLD_PRESS_TIME = 0.15
-export const DEFAULT_WINDOW_SECONDS = 0.15
-export const DEFAULT_STICK_DEADZONE_INNER = '0.15'
-export const DEFAULT_STICK_DEADZONE_OUTER = '0.10'
