--- src/constants/messages.ts
@@ -1,2 +1,0 @@
-export const LOCK_MESSAGE = 'Calibrating — place controller on a flat surface'
-export const PENDING_MESSAGE = 'Pending changes — click Apply to send to JoyShockMapper.'
